Co-authored-by: Brandon Do <brandonndo@users.noreply.github.com> Co-authored-by: Shi Kuang <shi-kuang@users.noreply.github.com> Co-authored-by: keniwane <keniwane@users.noreply.github.com> Co-authored-by: GMogi <GioMogi@users.noreply.github.com>
Refactor prototype-logic.js and add queryObjectFunctions.js

The refactor in prototype-logic.js includes several changes and adds the feature of extracting fragments definitions. This allows for better handling of object values. The code also adds a check for specific fields and sets certain fields to `true` in the prototype. Additionally, the code checks if the `proto` object has any `id` fields and sets `operationType` to "noID" if it does not.

The new queryObjectFunctions.js file includes function definitions for recreating the query for the database, toggling prototype fields to false, converting the query object to a string, and stringifying the query.

Co-authored-by: Brandon Do <brandonndo@users.noreply.github.com>
Co-authored-by: Shi Kuang <shi-kuang@users.noreply.github.com>
Co-authored-by: keniwane <keniwane@users.noreply.github.com>
Co-authored-by: GMogi <GioMogi@users.noreply.github.com>